test(tmm): pin TMM factors to real numeric oracles
